Continue Work on Collision System Improvements
Functional Build (except collision system sometimes causes 2 colliding objects to become perpetually stuck)

+++ src/main/java/com/taco/suit_lady/game/objects/tiles/GameTile.java
@@ -40,7 +40,7 @@ public class GameTile
     }
     
     public GameTile(@NotNull GameMap owner, @NotNull Dimensions dimensions) {
-        this(owner, dimensions.width(), dimensions.height());
+        this(owner, dimensions.width().intValue(), dimensions.height().intValue());
     }
     
     public GameTile(@NotNull GameMap owner, int xLoc, int yLoc) {
